Converted the app to Typescript
