+++ package.json
@@ -16,5 +16,8 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.8.0"
+  },
+  "devDependencies": {
+    "@types/react": "18.2.7"
   }
 }
