+++ pages/_app.tsx
@@ -1,0 +1,14 @@
+import 'bootstrap/dist/css/bootstrap.css';
+import '/styles/globals.css'
+import Layout from '../src/components/layout/Layout';
+import { AppProps } from 'next/app';
+
+export default function App({ Component, pageProps }: AppProps) {
+  return (
+    <>
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
+    </>
+  )
+}
