--- pages/index.js
@@ -1,41 +1,0 @@
-import Head from 'next/head'
-import { Inter } from 'next/font/google'
-import HomePage from '@/src/components/home/Home';
-
-
-const inter = Inter({ subsets: ['latin'] })
-
-export const getStaticProps = async () => {
-  const response = await fetch('https://jsonplaceholder.typicode.com/users');
-  const customers = await response.json();
-  return { 
-    props: { 
-      customers 
-    } 
-  };
-};
- 
-
-export default function Home({ customers }) {
-  return (
-    <>
-      <Head>
-        <title>Profile Navigator</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <HomePage customers={customers}/>
-      </>
-  )
-}
-
-// export async function getServerSideProps() {
-//   const response = await fetch('https://jsonplaceholder.typicode.com/users')
-//   const customers = await response.json()
-//   return {
-//     // pass the props you want to pass to your component above
-//     props: { customers }
-//   }
-// }
-
