--- src/components/navbar/Navbar.js
@@ -1,15 +1,0 @@
-import React from 'react'
-
-const Navbar = () => {
-  return (
-    <header>
-      <nav class="navbar  navbar-dark bg-primary fixed-top mb-5">
-        <div class="container-fluid">
-          <span class="navbar-brand mb-0 h1 fw-bolder">Profile Navigator</span>
-        </div>
-      </nav>
-    </header>
-  )
-}
-
-export default Navbar
