+++ src/components/navbar/Navbar.tsx
@@ -1,0 +1,15 @@
+import React from 'react'
+
+const Navbar = () => {
+  return (
+    <header>
+      <nav className="navbar  navbar-dark bg-primary fixed-top mb-5">
+        <div className="container-fluid">
+          <span className="navbar-brand mb-0 h1 fw-bolder">Profile Navigator</span>
+        </div>
+      </nav>
+    </header>
+  )
+}
+
+export default Navbar
