--- styles/Home.module.css
@@ -1,45 +1,0 @@
-/* .stickyHeader {
-  position: sticky;
-  top: 0;
-  background-color: #ffffff;
-  z-index: 100;
-}
-
-.title {
-  font-size: 2.6rem;
-  padding: 10px;
-  border-bottom: 3px solid #565252;
-  color: rgb(0, 72, 255);
-  text-shadow: 0 0 1px #000;
-}
-
-.grid__wrapper {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-template-rows: auto;
-  padding: .7em 3em;
-  gap: 20px;
-  margin-top: 2em;
-}
-
-.card {
-  background-color: #ccc;
-  display: block;
-  border-left: 9px solid #111111;
-  border-radius: 9px;
-  display: flex;
-  flex-direction: column;
-  padding: 1em 2em;
-  gap: 10px;
-
-}
-
-.card__title {
-  font-size: 1.1rem;
-}
-
-.card:hover,
-.card:focus {
-  border-left: 8px solid rgb(41, 183, 41);
-  cursor: pointer;
-} */
